backend/django/requirements.txt
django==3.0.8
djangorestframework==3.11.0
psycopg2-binary==2.8.5
pyyaml==5.3.1
uritemplate==3.0.1
hashids==1.3.1
django-cors-headers==3.4.0
djangorestframework-simplejwt==4.4.0